requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.3.0
murmurhash==1.0.2
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.2
packaging==20.4
Pillow==7.0.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
protobuf==3.11.3
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-Levenshtein==0.12.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
setuptools==40.0.0
six==1.14.0
sklearn==0.0
spacy==2.2.4
srsly==1.0.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tflearn==0.3.2
thinc==7.4.0
tqdm==4.45.0
typed-ast==1.4.1
urllib3==1.25.8
vaderSentiment==3.3.1
wasabi==0.6.0
wcwidth==0.2.3
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.1.0
